Refactor the trimmer code
I want to avoid a monolithic code for the trimmer, thus I made some decoupling here and there resulting new classes, e.g., Operation, PruneOperation, TargetFields, etc. I also created some static methods to make the trimmer easy to use and to read.

For example,
```
ObjectNode output =
  new JSONTreeTrimmer(input)
        .prune(at("p1", "p2"))
        .collapse(at("p3"))
        .trim();
```
Another reason for the refactoring is to prepare a conditional functionality to setup a condition that will ensure the pruning or collapsing only be executed if the condition is satisfied.