Create JsonLdMapper for local ObjectMapper use
This class is a copy of org.metadatacenter.util.json.JsonMapper that
will provide a static reference to Jackson ObjectMapper. The creation
of this static reference is safe and recommended according to:

http://stackoverflow.com/questions/3907929/should-i-declare-jacksons-obj
ectmapper-as-a-static-field

We didn’t reuse JsonMapper to avoid misaligned dependency, such that
the cedar-model-validation-library should not have any dependency to
cedar-server-core-library.